requirements.txt
click==8.1.3
pytest==7.4.0
chromadb==0.4.15
sentence-transformers==2.2.2
psycopg2-binary==2.9.1
python-dotenv==0.19.0